+++ src/receivers/wfmrx.cpp
@@ -149,7 +149,7 @@ void wfmrx::set_agc_target_level(int target_level)
     agc->set_target_level(target_level);
 }
 
-void wfmrx::set_agc_manual_gain(int gain)
+void wfmrx::set_agc_manual_gain(float gain)
 {
     agc->set_manual_gain(gain);
 }
